add im register center

+++ XSBK-im-load_balance/src/main/java/com/xsbk/balancer/common/model/ZkProperties.java
@@ -10,5 +10,5 @@ import lombok.Data;
 @Data
 public class ZkProperties {
 	private String serverAddress;
-	private long sessionTime;
+	private int sessionTime;
 }
